requirements.txt
requests
beautifulsoup4
pymupdf
pandas
sentence-transformers
faiss-cpu
python-dotenv
fastapi
uvicorn
